changed egauge usage and generation to bigint(20)

+++ schema.sql
@@ -60,8 +60,8 @@ CREATE TABLE `controleventresults` (
 CREATE TABLE `egauge` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `timestamp` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-  `usage` decimal(20, 4) NOT NULL,
-  `generation` decimal(20, 4) NOT NULL,
+  `usage` bigint(20) NOT NULL,
+  `generation` bigint(20) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
